Split regtest.go into focused files (same package)
Closes #38.

Phase 1.4 of the improvement roadmap. Pure mechanical split — no logic
changes, no API changes (all exports stay in package regtest).

Layout
- regtest.go (441 lines, was 1227): Config + Regtest types, New,
  DefaultConfig, RPCConfig, lifecycle (Start/Stop/Cleanup/IsRunning +
  Context variants, isConnRefusedErr), and internal helpers
  (initialize, extractPort, connectClient).
- rpc.go: Client(), GetBlockCount, HealthCheck (+ Context), and the
  rawRPC / lockedClient / runWithContext plumbing.
- wallet.go: GetWalletInformation, CreateWallet, LoadWallet,
  UnloadWallet, EnsureWallet (each with Context variant).
- address.go: GenerateBech32 / GenerateBech32m (+ Context) plus the
  shared generateAddress helper.
- mining.go: Warp / WarpContext.
- tx.go: ScantxoutsetUnspent / ScantxoutsetResult types,
  SendToAddress, GetTxOut, ScanTxOutSetForAddress,
  SignRawTransactionWithWallet, BroadcastTransaction (+ Context
  variants).

Verification
- go build ./... and go vet ./... clean.
- go doc -all . lists the same 41 exported funcs/methods as before.
- Full suite green under -race (count=1, ~65s).
